Continue with the transition to more strongly typed boxes
Get BoxTypesParsing to work on Sample.hs

Introduce an API based on the Gloss library's Picture and Event types,
which shouldn't be too hard to convert to js and SVGs if I want to do
that at some point.

`incDepthsExpr` may not be doing something sensible where it's used,
I've got it to typecheck without fully considering the contexts.
Nothing is testing it at the moment, so keep an eye out.